Canonicalize names/sources
This will try to infer the proper name for a module pulled from git
based on the repo it's pulled from. It will compare that repo against a
list of the `source` metadata from all Forge modules to identify what
the fully namespaced module name should be.

It will also better describe what programmatically calculated sources
are. Since we're doing static analysis, we can't interpolate a string
and have to just discard the value entirely.